backend/requirements.txt
aiosmtplib==4.0.2
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coverage==7.10.7
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lxml==6.0.2
lxml-stubs==0.5.1
markdown-it-py==4.0.0
mdurl==0.1.2
motor==3.7.1
mypy==1.14.1
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pyasn1==0.6.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.1
pytest==8.3.5
pytest-cov==5.0.0
pytest-env==1.1.5
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rich==14.1.0
rsa==4.9.1
ruff==0.13.1
selectolax==0.3.34
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
types-passlib==1.7.7.20250602
types-pyasn1==0.6.0.20250914
types-python-jose==3.5.0.20250531
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
